Adds cast of jsonpointer '-' part as list
